--- src/canisters/frontend/ic-certified-assets/src/certification/internals/mod.rs
@@ -1,3 +1,0 @@
-pub mod certification_types;
-pub mod http_types;
-pub mod tree;
