+++ src/canisters/frontend/ic-certified-assets/src/certification/types/mod.rs
@@ -1,0 +1,2 @@
+pub mod certification_types;
+pub mod http;
